+++ README.md
@@ -1,10 +1,7 @@
 # CoronaCare
-<p align = "center"> <image src="./CoronaCare/images/13coronavirus-explainer-mobileMasterAt3x-v2.jpg"> </p>
+<p align = "center"> <image src="./CoronaCare/src/assets/images/13coronavirus-explainer-mobileMasterAt3x-v2.jpg"> </p>
 
 ## What is CoronaCare?
 A mobile app platform to record and upload your daily temperature to myaces. 
 It makes surviging Covid-19 and daily temperature taking a breeze!
 
-## Testing
-
-HELLO WORLD
